META-124: Removed obsolete and incorrect comments; fixed a warning; added a missing include
<==== Mercurial data ====>
Branch: meta_4.0

+++ vms/libs/nx_sdk/src/nx/sdk/analytics/common/object_metadata_packet.cpp
@@ -33,7 +33,7 @@ int64_t ObjectMetadataPacket::durationUs() const
 
 IObject* ObjectMetadataPacket::nextItem()
 {
-    if (m_currentIndex < m_objects.size())
+    if (m_currentIndex < (int) m_objects.size())
         return m_objects[m_currentIndex++];
 
     return nullptr;
